+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.8.8**
 * It takes over 20s to bootstrap app in dev mode with large file in resources dir #846
+* FastJson exception when serialize ACT error message #841
 * `CacheFor` - make it allow skip `Cache-Control` header #837
 * `ResponseCache` - content disposition shall be cached #838
 * `CacheFor` key shall include `Accept` type #836